chat-app/pyproject.toml
[tool.poetry]
name = "chat-app"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.7.2"
chainlit = "^1.1.402"
python-dotenv = "^1.0.1"
google-cloud-aiplatform = "^1.63.0"
tenacity = "^9.0.0"
rich = "^13.7.1"
pillow = "^10.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
